feat: animal slider 리스트 렌더링
